Working camera

+++ 2D Minigames/src/game/engine/Time.java
@@ -1,8 +1,8 @@
 package game.engine;
 public class Time 
 {
-	public static float time;
-	public static float deltaTime;
-	public static float fixedTime;
-	public static float deltaFixedTime;
+	public static long time;
+	public static long deltaTime;
+	public static long fixedTime;
+	public static long deltaFixedTime;
 }
